Remove deprecated unload event triggering save
It’s not clear that this behaviour is desired since by now most actions
automatically save the app state. The only situation I can see this
being annoying is when you close the page without unfocusing the script
name or author input fields, in which case you’ll lose that info. I think
this is sufficiently rare not to be a problem.