Automatically make bloodstar night order explicit
Bloodstar night orders are a bit of a pain because you can’t put
anything before Minion/Demon info, before dusk, after dawn, etc. To fix
this you need to edit the JSON to add an explicit night order. This
update, combined with the previous, gets the bloodstar night order
almost right but allows you to edit the night order in the app, which is
then written to an explicit night order that you can use elsewhere.