Bring night reminder text in line with official
It’s not exactly the same as the usual night reminder text since in many
places it is inaccurate or very unclear.

The most significant change is the Damsel’s position in the night order.
The official reminder text for the Damsel instructs you to run the
Huntsman’s ability if the Huntsman chose the Damsel. This doesn’t make
sense – this is the Huntsman’s ability, not the Damsel’s (and the
Huntsman’s night reminder text already says this). To run the Damsel’s
ability, you have to inform the Minions that a Damsel is in play on the
first night right after Minion info. The night reminder text reflects
this.

Some reminder text is included that is omitted in the official tool. The
Cannibal and the Hermit have reminders to make sure you don’t forget
that they may have a waking ability. Riot reminder text ensures you
don’t forget to inform the Minions that they become Riot on day 3.