Implement toggleable autosorting + misc features
The script tool has always automatically sorted scripts. This is the
best experience by default for almost all scripts, which are in sort
order. It’s also the best experience for beginners, who may not even
know that scripts *have* a standard sort order. (Many scripts are posted
to the Blood on the Clocktower subreddit out of order.)

But it is also clearly inflexible: some scripts (like Ekin’s world cup
script Trained Killer) should be out of sort order, and the script tool
should allow people to work on such scripts. This is now possible.

A new checkbox is added to enable or disable autosorting and characters
can be dragged out of position to reorder them. The rules are
as follows:

- New scripts will by default automatically sort (‘autosort’) themselves.
- Imported scripts will autosort if the characters are already listed in
  sort order, otherwise they will not autosort. (This ensures that
  scripts like Ekin’s Trained Killer retain their custom ordering on
  import.)
- Turning autosort on manually will sort the script (this can be
  undone).
- Changes to the script that communicate intent to reorder characters
  manually will disable autosort. Undoing any such changes will
  re-enable autosort.

Some other things were changed:

- Bootlegger rules are now included in script links
- Minor undo and redo bugs fixed